Internal refactoring
All file/resource access/management utilities are now under the sub-packages ``stores``.
