+++ argopy/__init__.py
@@ -15,6 +15,7 @@ from . import tutorial
 
 # Other Import
 from . import utilities
+from . import stores
 from .utilities import show_versions
 from .options import set_options
 from .plotters import open_dashboard as dashboard
@@ -33,6 +34,7 @@ __all__ = (
     "utilities",
     "errors",
     "plotters",
+    "stores",
     # Constants
     "__version__"
 )
