+++ argopy/stores/__init__.py
@@ -1,0 +1,12 @@
+from .index import indexstore, index_filter_wmo
+from .fsspec_wrappers import filestore, ftpstore, httpstore
+
+#
+__all__ = (
+    # Classes:
+    "indexstore",
+    "index_filter_wmo",
+    "filestore",
+    "ftpstore",
+    "httpstore"
+)
